Job (spe v1.0.0)

Job module is responsible for managing the execution of tasks in a distributed system. It handles task assignment to workers, monitors their completion, and manages the overall state of the job. It uses a GenServer to maintain the state and handle asynchronous messages related to task execution and worker readiness.

Usage:

To use this module, you would typically start a job with Job.start_link/1 and then send tasks to it using Job.task_completed/2 and Job.worker_ready/2. It is designed to work with a distributed system where tasks can be executed by multiple workers, and it handles the complexities of task dependencies and worker availability.

{:ok, job_pid} = Job.start_link(%{id: "job_1", tasks: %{"task_1" => %{"exec" => "some_exec", "timeout" => 5000}}})
Job.worker_ready(job_pid, self())
Job.task_completed(job_pid, {"task_1", {:result, "task_result"}, self()})

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_cast(msg, state)

Handles asynchronous messages related to worker readiness and task completion. This function processes messages sent to the job, such as when a worker is ready to take tasks or when a task has been completed by a worker. It updates the job state accordingly and checks if the job should finish based on the current state.

Parameters:

message: The message to be processed, which can be either a worker readiness notification or a task completion notification.

state: The current state of the job, which includes ongoing tasks, free workers, and results.

Returns:

{:noreply, new_state}: Returns the updated state of the job wrapped in a :noreply tuple.

Example:

GenServer.cast(job_pid, {:worker_ready, worker_pid})
GenServer.cast(job_pid, {:task_terminated, {task_name, {:result, value}, worker_pid}})

handle_info(msg, state)

Handles informational messages related to worker readiness notifications. This function processes messages sent to the job when a worker is ready to take tasks. It updates the job state by adding the worker to the list of free workers and logs the addition of the new worker.

Parameters:

message: The informational message containing the worker PID.

state: The current state of the job, which includes ongoing tasks, free workers, and results.

Returns:

{:noreply, state}: Returns the current state of the job wrapped in a :noreply tuple.

Example:

GenServer.handle_info({:notify_ready, worker_pid}, state)

init(state)

Initializes the job state when the GenServer starts. This function sets up the initial state of the job, including the task returns, ongoing tasks, free workers, results, and the start time. It also logs the start of the job for debugging purposes.

Parameters:

state: The initial state of the job, which includes the job ID and tasks to be executed.

Returns:

{:ok, state}: Returns the initial state of the job wrapped in an :ok tuple.

Example:

{:ok, initial_state} = Job.init(%{id: "job_1", tasks: %{"task_1" => %{"exec" => "some_exec", "timeout" => 5000}}})

start_link(state)

Starts the job with the given initial state. This function initializes the job by setting up the necessary state and prepares it to handle tasks and workers. It also logs the start of the

Parameters:

state: The initial state of the job, which includes the job ID and tasks to be executed. Returns::

{:ok, pid}: Returns the PID of the started job process wrapped in an :ok tuple. mple: :
{:ok, job_pid} = Job.start_link(%{id: "job_1", tasks: %{"task_1" => %{"exec" => "some_exec", "timeout" => 5000}
}})

task_completed(job_id, arg)

Notifies the job that a task has been completed by a worker. This function is called when a worker finishes executing a task. It sends a message to the job process indicating that the task has been completed, along with the task name, result, and worker PID. The job process will then handle this message to update its state accordingly.

Parameters:

job_id: The ID of the job process that is handling the task.

task_name: The name of the task that has been completed.

result: The result of the task execution, which can be a success or failure.

worker_pid: The PID of the worker that completed the task.

Example:

Job.task_completed(job_id, {"task_1", {:result, "task_result"}, worker_pid})

worker_ready(job_pid, worker_pid)

Notifies the job that a worker is ready to take tasks. This function is called when a worker signals that it is ready to receive tasks. It sends a message to the job process indicating that the worker is ready, allowing the job to assign tasks to the worker as they become available.

Parameters:

job_pid: The PID of the job process that is managing the workers.

worker_pid: The PID of the worker that is ready to take tasks.

Example:

Job.worker_ready(job_pid, worker_pid)