JobManager (spe v1.0.0)

JobManager is responsible for managing jobs, including submitting jobs, starting them, and handling their plans. It uses a GenServer to maintain state and handle calls related to job management.

Usage:

Start the JobManager
{:ok, _pid} = JobManager.start_link(num_workers: 4)

Submit a job
job_desc = %{
"tasks" => [
  %{"name" => "task1", "enables" => ["task2"]},
  %{"name" => "task2", "enables" => []}
]
}
{:ok, job_id} = JobManager.submit_job(job_desc)

Start the job
{:ok, _job_id} = JobManager.start_job(job_id)

Notes:

The JobManager uses a planning module (Planner) to create a plan for the job based on the job description and the number of workers available.

It handles job submissions and starts jobs based on the plans created.

The JobManager maintains a state that includes the jobs and their plans, as well as the number of workers available for job execution.

It logs debug information about job submissions, plans, and task management.

Dependencies:

GenServer: For managing the state and handling calls.

Logger: For logging debug information.

Planner: A module responsible for planning jobs based on their descriptions and available workers.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_call(msg, from, state)

Handles job submission by receiving a job description, planning the job, and storing it in the state.

Parameters:

{:submit, job_desc}: A tuple containing the job description to be submitted.

Returns:

{:reply, {:ok, job_id}, new_state}: If the job is successfully planned and stored, it returns the job ID and the updated state.

{:reply, {:error, error}, state}: If there is an error in planning the job, it returns the error and the current state.

Example:

job_desc = %{
  "tasks" => [
    %{"name" => "task1", "enables" => ["task2"]},
    %{"name" => "task2", "enables" => []}
  ]
}
{:ok, job_id} = JobManager.submit_job(job_desc)

handle_info(msg, state)

Handles job planning by receiving a job ID and its planned tasks, updating the state with the new plan.

Parameters:

{:planning, {job_id, job_plan}}: A tuple containing the job ID and its planned tasks.

Returns:

{:noreply, new_state}: The updated state after planning the job.

Example:

job_id = "12345"
job_plan = [%{"task" => "task1"}, %{"task" => "task2"}]
JobManager.plan_ready(job_id, job_plan)

init(state)

Initializes the JobManager with a state containing the number of workers, jobs, and waiting jobs. The num_workers can be set to :unbound to allow dynamic worker allocation based on job requirements. The jobs map will hold the jobs submitted, and waiting can be used for jobs that are waiting to be processed.

Parameters:

state: The initial state of the JobManager, including the number of workers, jobs, and waiting jobs.

Returns:

{:ok, state}: The initial state of the JobManager is returned, ready to handle job submissions and starts.

Example:

{:ok, _pid} = JobManager.start_link(num_workers: 4)

plan_ready(job_id, job_plan)

Notifies the JobManager that a job plan is ready for the given job ID.

Parameters:

job_id: The ID of the job for which the plan is ready.

job_plan: The planned tasks for the job.

Returns:

:ok: If the plan is successfully received and processed.

Example:

job_id = "12345"
job_plan = [%{"task" => "task1"}, %{"task" => "task2"}]
JobManager.plan_ready(job_id, job_plan)

start_job(job_id)

Starts a job with the given job ID.

Parameters:

job_id: The ID of the job to be started.

Returns:

{:ok, job_id}: If the job is successfully started, it returns the job ID.

{:error, error}: If there is an error in starting the job, it returns the error.

Example:

{:ok, _job_id} = JobManager.start_job("12345")

start_link(opts)

Starts the JobManager GenServer with the specified options.

Parameters:

opts: A keyword list of options, including :num_workers to specify the number of workers and :name to set the GenServer name.

Returns:

{:ok, pid}: The PID of the started JobManager GenServer.

Example:

{:ok, _pid} = JobManager.start_link(num_workers: 4, name: :job_manager)

submit_job(job_desc)

Submits a job to the JobManager with the given job description.

Parameters:

job_desc: A map containing the job description, including tasks and their dependencies.

Returns:

{:ok, job_id}: If the job is successfully submitted and planned, it returns the job ID.

{:error, error}: If there is an error in planning the job, it returns the error.

Example:

job_desc = %{
  "tasks" => [
    %{"name" => "task1", "enables" => ["task2"]},
    %{"name" => "task2", "enables" => []}
  ]
}
{:ok, job_id} = JobManager.submit_job(job_desc)