Planner (spe v1.0.0)

Planner is responsible for planning jobs based on their descriptions and the number of workers available. It implements Kahn's algorithm to create a plan for executing tasks while respecting their dependencies.

Usage:

Call Planner.planning/2 with a job description and the number of workers to get a planned execution order.

Example:

job_desc = %{
  "name" => "nisse",
  "tasks" => [
    %{
      "name" => "t0",
      "enables" => [],
      "exec" => fn _ -> 1 + 2 end,
      "timeout" => :infinity
    },
    %{
      "name" => "t1",
      "enables" => ["t0"],
      "exec" => fn _ -> 3 + 4 end,
      "timeout" => :infinity
    }
  ]
}
{:ok, plan} = Planner.planning(job_desc, 2)

Notes:

The job description should contain a list of tasks, each with a name and a list of tasks it enables.

The planning/2 function will return an ordered list of task groups that can be executed in parallel, respecting their dependencies.

If the job description contains a "priority" field, it will be used to prioritize tasks in the planning.

Functions

find_dependent_tasks(enables, start_task)

Finds all tasks that depend on a given start task, including transitive dependencies.

Parameters:

enables: A map where keys are task names and values are lists of tasks that the key task enables.

start_task: The task from which to start finding dependent tasks.

Returns:

A list of all tasks that depend on the start_task, including transitive dependencies.

Example:

enables = %{
  "task1" => ["task2", "task3"],
  "task2" => ["task4"],
  "task3" => [],
  "task4" => []
}
start_task = "task1"
dependent_tasks = Planner.find_dependent_tasks(enables, start_task)

find_next_independent(list, deps, done)

find_next_independent(list, deps, undone, done)

Finds the next independent task from a list of tasks, given their dependencies and the already completed tasks.

Parameters:

tasks: A list of tasks to check for independence.

deps: A map where keys are task names and values are lists of dependencies for each task.

undone: A list of tasks that are yet to be completed.

done: A list of tasks that have already been completed.

Returns:

The first independent task found in the list, or nil if no independent task is found.

Example:

tasks = ["task1", "task2", "task3"]
deps = %{"task1" => [], "task2" => ["task1"], "task3" => ["task1", "task2"]}
undone = ["task2", "task3"]
done = ["task1"]
next_task = Planner.find_next_independent(tasks, deps, undone, done)

group_tasks(tasks, deps, num_workers)

Groups tasks into independent sets, considering task dependencies and worker limits.

Parameters:

tasks: A list of tasks to be grouped.

deps: A map where keys are task names and values are lists of dependencies for each task.

num_workers: The maximum number of workers available for executing tasks. If :unbound, it allows dynamic allocation.

Returns:

A list of lists, where each inner list contains tasks that can be executed in parallel, respecting their dependencies.

Example:

tasks = ["task1", "task2", "task3"]
deps = %{"task1" => [], "task2" => ["task1"], "task3" => ["task1", "task2"]}
num_workers = 2
grouped_tasks = Planner.group_tasks(tasks, deps, num_workers)

group_tasks_with_priority(tasks, deps, num_workers, priority)

Groups tasks into independent sets, considering task dependencies and worker limits.

Parameters:

tasks: A list of tasks to be grouped.

deps: A map where keys are task names and values are lists of dependencies for each task.

num_workers: The maximum number of workers available for executing tasks. If :unbound, it allows dynamic allocation.

priority: A list of tasks that should be prioritized in the grouping.

Returns:

A list of lists, where each inner list contains tasks that can be executed in parallel, respecting their dependencies and priorities.

Example:

tasks = ["task1", "task2", "task3"]
deps = %{"task1" => [], "task2" => ["task1"], "task3" => ["task1", "task2"]}
num_workers = 2
priority = ["task1"]
grouped_tasks = Planner.group_tasks_with_priority(tasks, deps, num_workers, priority)

khan_loop(dependencies, free_tasks, planned)

Implements Kahn's algorithm to find a topological ordering of tasks based on their dependencies.

Parameters:

dependencies: A map where keys are task names and values are lists of tasks that depend on the key task.

free_tasks: A list of tasks that have no dependencies and can be executed immediately.

planned: A list that accumulates the planned tasks in the order they can be executed.

Returns:

{:ok, planned}: If a valid topological ordering is found, returning the ordered list of tasks.

{:error, :graph_has_cycle}: If a cycle is detected in the task dependencies, making it impossible to create a valid ordering.

Example:

dependencies = %{
  "task1" => ["task2"],
  "task2" => [],
  "task3" => ["task1"]
}
free_tasks = ["task2"]
planned = []
{:ok, ordered_tasks} = Planner.khan_loop(dependencies, free_tasks, planned)

planning(job_desc, num_workers)

Plans a job based on its description and the number of workers available. It uses Kahn's algorithm to create a plan for executing tasks while respecting their dependencies.

Parameters:

job_desc: A map containing the job description, including tasks and their dependencies.

num_workers: The number of workers available for executing the tasks. It can be an integer or :unbound for dynamic allocation.

Returns:

{:ok, plan}: A list of task groups that can be executed in parallel, respecting their dependencies.

{:error, reason}: If the job description is invalid or if there is a cycle in the task dependencies.

Example:

job_desc = %{
  "name" => "nisse",
  "tasks" => [
    %{
      "name" => "t0",
      "enables" => [],
      "exec" => fn _ -> 1 + 2 end,
      "timeout" => :infinity
    },
    %{
      "name" => "t1",
      "enables" => ["t0"],
      "exec" => fn _ -> 3 + 4 end,
      "timeout" => :infinity
    }
  ]
}
{:ok, plan} = Planner.planning(job_desc, 2)