SPE (spe v1.0.0)

SPE is the main module for the SPE (Scalable Parallel Execution) system. It is responsible for starting the server, managing job submissions, and handling job starts. It uses a GenServer to maintain the state of the system and handle calls related to job management.

Usage:

Start the SPE server
{:ok, _pid} = SPE.start_link(num_workers: 4)

Submit a job
job_desc = %{
"tasks" => [
  %{"name" => "task1", "enables" => ["task2"]},
  %{"name" => "task2", "enables" => []}
]
}
{:ok, job_id} = SPE.submit_job(job_desc)

Start the job
{:ok, _job_id} = SPE.start_job(job_id)

Notes:

The SPE module uses a JobManager to handle job submissions and starts.

It maintains a state that includes waiting jobs, which are jobs that are waiting for their plans to be ready before they can start.

It logs debug information about job submissions, starts, and other events.

Dependencies:

GenServer: For managing the state and handling calls.

Logger: For logging debug information.

JobManager: A module responsible for managing jobs, including submitting jobs, starting them, and handling their plans.

SuperManager: A module responsible for managing the overall system and ensuring that the SPE server is running.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_call(msg, from, state)

Handles synchronous calls to the SPE server. It processes job start requests and job submissions.

Parameters:

{:start_job, job_id}: A tuple containing the job ID to start.

{:submit_job, job_desc}: A tuple containing the job description to submit.

Returns:

{:reply, response, state}: The response from the JobManager for the job start or submission, along with the current state.

Example:

{:ok, job_id} = SPE.start_job("job_123")
{:ok, job_id} = SPE.submit_job(%{"tasks" => [%{"name" => "task1", "enables" => []}]})

handle_info(msg, state)

Handles generic info messages sent to the SPE server. It logs the received message and returns the current state without any changes.

Parameters:

msg: The message received by the SPE server.

Returns:

{:noreply, state}: The current state of the SPE server is returned without any changes.

Example:

SPE.handle_info(:some_info_message, state)

init(arg)

Initializes the SPE server with the given options and state. The options can include the number of workers, which will be passed to the JobManager.

Parameters:

opts: A keyword list of options, including :num_workers to specify the number of workers.

state: The initial state of the SPE server, which includes a map for waiting jobs.

Returns:

{:ok, state}: The initial state of the SPE server is returned, ready to handle job submissions and starts.

Example:

{:ok, _pid} = SPE.start_link(num_workers: 4)

job_ready(job_id, response)

Notifies the SPE server that a job is ready to start. This function is called when a job plan is ready, and it will notify the waiting clients that the job can now start.

Parameters:

job_id: The ID of the job that is ready to start.

response: The response from the JobManager indicating whether the job can start or if there was an error.

Returns:

:ok: If the job is successfully marked as ready to start.

Example:

job_id = "12345"
response = {:ok, job_id}
SPE.job_ready(job_id, response)

start_job(job_id)

Starts a job with the given job ID.

Parameters:

job_id: The ID of the job to be started.

Returns:

{:ok, job_id}: If the job is successfully started, it returns the job ID.

{:error, error}: If there is an error in starting the job, it returns the error.

Example:

{:ok, _job_id} = SPE.start_job("12345")

start_link(opts)

Starts the SPE server with the given options.

Parameters:

opts: A keyword list of options, including :num_workers to specify the number of workers.

Returns:

{:ok, pid}: The PID of the started SPE server.

Example:

{:ok, _pid} = SPE.start_link(num_workers: 4)

submit_job(job_desc)

Submits a job to the SPE server with the given job description.

Parameters:

job_desc: A map containing the job description, including tasks and their dependencies.

Returns:

{:ok, job_id}: If the job is successfully submitted and planned, it returns the job ID.

{:error, :invalid_description}: If the job description is invalid, it returns an error.

Example:

job_desc = %{
  "tasks" => [
    %{"name" => "task1", "enables" => ["task2"]},
    %{"name" => "task2", "enables" => []}
  ]
}
{:ok, job_id} = SPE.submit_job(job_desc)

terminate(reason, state)

Terminates the SPE server and stops the Supervisor managing it. This function is called when the SPE server is no longer needed, and it will clean up the resources used by the server.

Parameters:

_reason: The reason for termination, which is not used in this implementation.

state: The current state of the SPE server, which includes the Supervisor's PID.

Returns:

:ok: Indicates that the SPE server has been successfully terminated.

Example:

SPE.terminate(:normal, state)