SuperManager (spe v1.0.0)

SuperManager is responsible for managing the lifecycle of jobs in the SPE system. It uses a Supervisor to manage the SuperJob processes, which handle the execution of jobs and their associated workers.

Usage:

Start the SuperManager
{:ok, _pid} = SuperManager.start_link()

Start a job with the given job state and number of workers
job_state = %{id: "job_1", ...} # Define the job state as needed
{:ok, _job_pid} = SuperManager.start_job(job_state, 4)

Notes:

The SuperManager uses a one-for-one restart strategy, meaning that if a job fails, it will be restarted individually.

It logs debug information about job starts and the SuperJob processes.

Dependencies:

Supervisor: For managing the SuperJob processes.

Logger: For logging debug information.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

init(init_arg)

Initializes the Supervisor with the necessary options. This includes setting the strategy to :one_for_one and defining the maximum number of restarts and seconds.

Parameters:

_init_arg: Initialization argument, not used in this case.

Returns:

{:ok, pid}: The Supervisor's PID is returned, indicating that it has been successfully started.

Example:

{:ok, sup_pid} = SuperManager.start_link()

start_job(job_state, num_workers)

Starts a SuperJob with the given job state and number of workers. This function initializes the SuperJob process and starts it under the SuperManager Supervisor.

Parameters:

job_state: A map containing the job state, including the job ID and any other necessary information.

num_workers: The number of workers to be started for the job.

Returns:

{:ok, pid}: The PID of the started SuperJob process is returned, indicating that the job has been successfully started.

Example:

job_state = %{id: "job_1", ...} # Define the job state as needed
{:ok, _job_pid} = SuperManager.start_job(job_state, 4)

start_link()

Starts the SuperManager Supervisor. This function initializes the Supervisor and sets it up to manage SuperJob processes.

Returns:

{:ok, pid}: The Supervisor's PID is returned, indicating that it has been successfully started.

Example:

{:ok, sup_pid} = SuperManager.start_link()