SuperSPE (spe v1.0.0)

SuperSPE is the main supervisor for the SuperSPE application. It manages the SuperManager and JobManager processes.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

init(children)

Starts the SuperSPE supervisor with its children processes. The children include the Phoenix PubSub, SuperManager, and JobManager. The supervisor uses a one-for-one strategy with a maximum of one restart within five seconds. The JobManager is started with a name specified in the options. The SuperManager is started without any options. The PubSub is started with a name defined in the SPE module. The supervisor is named SPE.SuperSPE.

Returns:

{:ok, pid}: The Supervisor's PID is returned, indicating that it has been successfully started.

Example:

{:ok, sup_pid} = SuperSPE.start_link([])

start_link(opts)

Starts the SuperSPE supervisor with the given options. This function initializes the Supervisor and sets it up to manage the SuperManager and JobManager processes.

Parameters:

opts: Options for the JobManager, including the name to be used.

Returns:

{:ok, pid}: The Supervisor's PID is returned, indicating that it has been successfully started.

Example:

{:ok, sup_pid} = SuperSPE.start_link(name: SPE.JobManager)