SuperWorker (spe v1.0.0)

SuperWorker is responsible for managing a set of workers that execute jobs in parallel. It uses a Supervisor to manage the workers, ensuring that they are restarted if they fail.

Usage:

Start a SuperWorker
{:ok, sup_pid} = SuperWorker.start_link(job_pid, num_workers)

The job_pid is the PID of the job that the workers will execute.

The num_workers specifies how many workers should be started for the job.

Notes:

The SuperWorker module uses a Supervisor to manage the workers.

It defines a one-for-one restart strategy, meaning that if a worker fails, it will be restarted individually.

It logs debug information about the worker startup processes.

Dependencies:

Supervisor: For managing the workers.

Logger: For logging debug information.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

init(info)

Initializes the Supervisor with the necessary options and child processes. This includes setting the strategy to :one_for_one and defining the maximum number of restarts and seconds.

Parameters:

info: A map containing the job PID and the number of workers to be started.

Returns:

{:ok, pid}: The Supervisor's PID is returned, indicating that it has been successfully started.

Example:

{:ok, sup_pid} = SuperWorker.start_link(job_pid, num_workers)

start_link(job_pid, num_workers)

Starts the SuperWorker Supervisor with the given job PID and number of workers. This function initializes the Supervisor and sets it up to manage the workers for the specified job.

Parameters:

job_pid: The PID of the job that the workers will execute.

num_workers: The number of workers to be started for the job.

Returns:

{:ok, pid}: The Supervisor's PID is returned, indicating that it has been successfully started.

Example:

{:ok, sup_pid} = SuperWorker.start_link(job_pid, num_workers)