Worker (spe v1.0.0)

Worker is responsible for executing tasks within a job in the SPE system. It is a GenServer that handles task execution, including applying functions with parameters and managing timeouts.

Usage:

Start a Worker
{:ok, worker_pid} = Worker.start_link(%{job: job_pid})

Send a task to the Worker
Worker.send_task(worker_pid, job_id, task_name, timeout, function, params)

Check if the Worker is ready
Worker.are_u_ready?(worker_pid)

Notes:

The Worker uses a GenServer to manage its state and handle task execution.

It applies functions with parameters and handles exceptions that may occur during execution.

It communicates task completion back to the Job that it belongs to.

Dependencies:

GenServer: For managing the state and handling calls.

Logger: For logging debug information.

Phoenix.PubSub: For broadcasting task start and termination events.

Functions

apply(job_id, task_name, timeout, function, args)

Applies a function with the given parameters and handles exceptions. This function executes the provided function with the specified parameters and manages timeouts and exceptions.

Parameters:

job_id: The ID of the job that the task belongs to.

task_name: The name of the task to be executed.

timeout: The timeout for the task execution, in milliseconds.

function: The function to be executed as part of the task.

args: The parameters to be passed to the function.

Returns:

{:result, result}: If the function executes successfully, it returns the result wrapped in a tuple.

{:failed, reason}: If the function fails or times out, it returns an error tuple with the reason.

Example:

result = Worker.apply(job_id, task_name, timeout, function, args)

are_u_ready?(worker_pid)

Checks if the Worker is ready to handle tasks. This function sends a call to the Worker to check its readiness.

Parameters:

worker_pid: The PID of the Worker to check.

Returns:

:ok: Indicates that the Worker is ready to handle tasks.

Example:

Worker.are_u_ready?(worker_pid)

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

handle_call(msg, from, state)

Handles the call to check if the Worker is ready. This function responds to a call to check if the Worker is ready to handle tasks.

Parameters:

:are_u_ready: A message indicating that the caller wants to check if the Worker is ready.

Returns:

{:reply, :ok, state}: Indicates that the Worker is ready and returns the current state.

Example:

GenServer.call(worker_pid, :are_u_ready)

handle_cast(msg, state)

Handles the task execution in the Worker. This function is called when a task is sent to the Worker. It applies the function with the given parameters and handles the result.

Parameters:

{:task, {job_id, {name, timeout, fun, params}}}: A tuple containing the job ID, task name, timeout, function to execute, and parameters.

Returns:

{:noreply, state}: Indicates that the Worker has completed the task and is ready for the next one.

Example:

GenServer.cast(worker_pid, {:task, {job_id, {task_name, timeout, function, params}}})

init(state)

Initializes the Worker with the given state. The state should include the job PID that the Worker is associated with.

Parameters:

state: A map containing the job PID and any other necessary state information.

Returns:

{:ok, state}: The initial state of the Worker is returned, ready to handle tasks.

Example:

{:ok, worker_pid} = Worker.start_link(%{job: job_pid})

send_task(worker_pid, job_id, name, timeout, fun, params)

Sends a task to the Worker for execution. This function sends a task to the Worker, which includes the job ID, task name, timeout, function to execute, and parameters.

Parameters:

worker_pid: The PID of the Worker to which the task will be sent.

job_id: The ID of the job that the task belongs to.

name: The name of the task to be executed.

timeout: The timeout for the task execution, in milliseconds.

fun: The function to be executed as part of the task.

params: The parameters to be passed to the function.

Returns:

:ok: Indicates that the task has been sent to the Worker for execution.

Example:

Worker.send_task(worker_pid, job_id, task_name, timeout, function, params)

start_link(state)

Starts the Worker process with the given state. This function initializes the Worker and sets it up to handle tasks.

Parameters:

state: A map containing the job PID and any other necessary state information.

Returns:

{:ok, pid}: The Worker process's PID is returned, indicating that it has been successfully started.

Example:

{:ok, worker_pid} = Worker.start_link(%{job: job_pid})